all warning resolution

+++ module/fiber.c
@@ -105,7 +105,7 @@ static long hit_ioctl(struct file *filp, unsigned int cmd, unsigned long __user 
         return -1;
       }
       copy_from_user(index,(long *)ptr, sizeof(long));
-      return (long) fls_get_value(*index);
+      return fls_get_value(*index);
 
     case IOCTL_FLS_SET:
       fls = kmalloc(sizeof(struct fls_info),__GFP_HIGH);
